+++ pennylane_lightning_gpu/src/util/MPIManager.hpp
@@ -30,7 +30,14 @@
 #include <unordered_map>
 #include <vector>
 
+#include "DataBuffer.hpp"
 #include "Error.hpp"
+
+/// @cond DEV
+namespace {
+using namespace Pennylane::CUDA;
+} // namespace
+/// @endcond
 
 namespace Pennylane::MPI {
 inline void errhandler(int errcode, const char *str) {
@@ -405,6 +412,15 @@ class MPIManager final {
         return recvBuf;
     }
 
+    /**
+     * @brief MPI_Reduce wrapper.
+     *
+     * @tparam T C++ data type.
+     * @param sendBuf Send buffer.
+     * @param recvBuf Receive buffer.
+     * @param root Rank of root process.
+     * @param op_str String of MPI_Op.
+     */
     template <typename T>
     void Reduce(T &sendBuf, T &recvBuf, size_t root,
                 const std::string &op_str) {
@@ -414,6 +430,15 @@ class MPIManager final {
                                      this->getComm()));
     }
 
+    /**
+     * @brief MPI_Reduce wrapper.
+     *
+     * @tparam T C++ data type.
+     * @param sendBuf Send buffer vector.
+     * @param recvBuf Receive buffer vector.
+     * @param root Rank of root process.
+     * @param op_str String of MPI_Op.
+     */
     template <typename T>
     void Reduce(std::vector<T> &sendBuf, std::vector<T> &recvBuf, size_t root,
                 const std::string &op_str) {
@@ -424,6 +449,33 @@ class MPIManager final {
                                      this->getComm()));
     }
 
+    /**
+     * @brief MPI_Reduce wrapper.
+     *
+     * @tparam T C++ data type.
+     * @param sendBuf Send buffer (DataBuffer type).
+     * @param recvBuf Receive buffer (DataBuffer type).
+     * @param root Rank of root process.
+     * @param op_str String of MPI_Op.
+     */
+    template <typename T>
+    void Reduce(DataBuffer<T> &sendBuf, DataBuffer<T> &recvBuf, size_t length,
+                size_t root, const std::string &op_str) {
+        MPI_Datatype datatype = getMPIDatatype<T>();
+        MPI_Op op = getMPIOpType(op_str);
+        PL_MPI_IS_SUCCESS(MPI_Reduce(sendBuf.getData(), recvBuf.getData(),
+                                     length, datatype, op, root,
+                                     this->getComm()));
+    }
+
+    /**
+     * @brief MPI_Reduce wrapper.
+     *
+     * @tparam T C++ data type.
+     * @param sendBuf Send buffer.
+     * @param recvBuf Receive buffer vector.
+     * @param root Rank of root process.
+     */
     template <typename T>
     void Gather(T &sendBuf, std::vector<T> &recvBuf, size_t root) {
         MPI_Datatype datatype = getMPIDatatype<T>();
@@ -431,6 +483,14 @@ class MPIManager final {
                                      datatype, root, this->getComm()));
     }
 
+    /**
+     * @brief MPI_Reduce wrapper.
+     *
+     * @tparam T C++ data type.
+     * @param sendBuf Send buffer vector.
+     * @param recvBuf Receive buffer vector.
+     * @param root Rank of root process.
+     */
     template <typename T>
     void Gather(std::vector<T> &sendBuf, std::vector<T> &recvBuf, size_t root) {
         MPI_Datatype datatype = getMPIDatatype<T>();
@@ -531,6 +591,39 @@ class MPIManager final {
                                       recvBuf.data(), recvBuf.size(), datatype,
                                       rootInt, this->getComm()));
         return recvBuf;
+    }
+
+    /**
+     * @brief MPI_Send wrapper.
+     *
+     * @tparam T C++ data type.
+     * @param sendBuf Send buffer vector.
+     * @param dest Rank of send dest.
+     */
+    template <typename T> void Send(std::vector<T> &sendBuf, size_t dest) {
+        MPI_Datatype datatype = getMPIDatatype<T>();
+        const int tag = 6789;
+
+        PL_MPI_IS_SUCCESS(MPI_Send(sendBuf.data(), sendBuf.size(), datatype,
+                                   static_cast<int>(dest), tag,
+                                   this->getComm()));
+    }
+
+    /**
+     * @brief MPI_Recv wrapper.
+     *
+     * @tparam T C++ data type.
+     * @param recvBuf Recv buffer vector.
+     * @param source Rank of data source.
+     */
+    template <typename T> void Recv(std::vector<T> &recvBuf, size_t source) {
+        MPI_Datatype datatype = getMPIDatatype<T>();
+        MPI_Status status;
+        const int tag = MPI_ANY_TAG;
+
+        PL_MPI_IS_SUCCESS(MPI_Recv(recvBuf.data(), recvBuf.size(), datatype,
+                                   static_cast<int>(source), tag,
+                                   this->getComm(), &status));
     }
 
     /**