+++ .github/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### New features since last release
 
+ * Add Sparse Hamiltonian support for expectation value calculation.
+ [(#127)] (https://github.com/PennyLaneAI/pennylane-lightning-gpu/pull/127)
+
+
 ### Breaking changes
 
 ### Improvements
@@ -13,6 +17,8 @@
 ### Contributors
 
 This release contains contributions from (in alphabetical order):
+
+Shuli Shu
 
 ---